feature createExam , updateExam , updatePublicExam , deleteExam & get single vs list exam

+++ backend-springboot/src/main/java/com/nlu/repository/ExamNumberRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nlu.model.entity.ExamNumber;
 
-public interface ExamNumberRepository extends JpaRepository<ExamNumber, Long> {
+public interface ExamNumberRepository extends JpaRepository<ExamNumber, Integer> {
 
 }
